% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_satchel_dir}
\alias{init_satchel_dir}
\title{initialize a satchel directory}
\usage{
init_satchel_dir(.dir = "data/derived/satchel")
}
\arguments{
\item{.dir}{directory}
}
\value{
path to satchel directory
}
\description{
initialize a satchel directory
}
\details{
Defaults to data/derived/satchel relative
to working directory

Best if run from the console after opening an rstudio project
}
