% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Satchel.R
\docType{class}
\name{Satchel}
\alias{Satchel}
\title{Carry around your data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Satchel
}
\description{
Carry around your data
}
\section{Initialization}{


\code{satchel <- Satchel$new("<namespace>", "<path/to/satchel>")}
\itemize{
\item The namespace will be the name of the folder/namespace that will be used to store
the data elements
\item the path must already exist, this is to protect random satchel directories from
being created in the case that the wrong directory path is set
}
}

\section{Methods}{


methods:
\itemize{
\item save(data, data_name) - save data
\item use(data_name, namespace) - use data saved from a different location
\item report() - show information about all data saved in current session
\item details() - list all data in the satchel cache directory
\item preview(data_name, namespace) - shows the (approximate) head of a dataset stored in the output metadata
}
}

\examples{
\dontrun{
# create a new satchel stored as namespace f1 in the dir data/derived
satchel <- Satchel$new("f1", "../data/derived/satchel")
satchel$save(Theoph)
# new file at ../data/derived/satchel/f1/Theoph.rds

# we can also give the files other names
satchel$save(Theoph, data_name = "other")
satchel$save(Theoph, data_name = "another")

# to see all objects saved during the session can check the report
satchel$report()

# can see data from any satchel dir by checking what is available
satchel$available()

# lets say in another file I had saved a file nca_summaries in an nca_analysis namespace
# satchel is smart enough to scan all folders for that file
nca_summaries <- satchel$use("nca_summaries")

# if it finds a conflict, you must explicitly specify which namespace it was saved under
nca_summaries <- satchel$use("nca_summaries", "nca_analysis")


}
}
\keyword{datasets}
